Adds Extension Tag for drawio
The existing extension is using parser-function {{drawio}} which is meant for template/translcusion or other in-page content manipulation.

This adds the extension tag support with <drawio></drawio> or preferably <drawio/> with all the same attributes as the original extension.

Both continue to be usable, but I would suggest promoting the extension tag as this is the intended tag format when adding extended capabilities to MW.
see: https://www.mediawiki.org/wiki/Manual:Tag_extensions

It also brings added benefit that Visual Editor can be extended easily to support this extension.

Bug: T271615
Change-Id: I8e59cd344a636632ceea6c806451225b00166854